reorganize page.tsx, move evalSocialData to utils
